pyproject.toml
[project]
name = "grok-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyinstaller>=6.13.0",
    "pywebview",
]
